add grid split column layout
